Round 8: tolerant Trigger int64s, update_mask honesty, hygiene
- serde_util.rs: lift serialize/deserialize_string_i64 out of
  environments.rs into a shared module; Trigger's three int64 fields now
  tolerate the protobuf-JSON string form live-verified on the sibling
  environments resource, so one string-encoded int degrades per-field
  instead of failing the whole TriggerListResponse (unit test pins it);
  serialization stays numeric — protobuf-JSON accepts both on input
- update_trigger: document that the SDK spec exposes no update_mask for
  triggers (unlike webhooks) and soften the partial-update claims on
  TriggerUpdate, the wire test, and the guide accordingly — field
  omission is the only scoping mechanism the wire offers, and the
  semantics stay unconfirmed while creation is agent-gated
- antigravity: replace the clone-then-expect with the let-chain borrow
  (step is an owned local, so no borrow conflict existed)
- tests: create_trigger is deliberately not retry-wrapped (a lost-
  response retry could leave a second scheduled trigger with no ID to
  clean up); the environments create retry is kept and the asymmetry
  documented (containers expire on their own, bounding the leak)
