+++ src/environments.rs
@@ -14,6 +14,7 @@
 //! convention); both are accepted here as numbers too.
 
 use crate::environment::{EnvironmentSource, NetworkConfig};
+use crate::serde_util::{deserialize_string_i64, serialize_string_i64};
 use chrono::{DateTime, Utc};
 use serde::de::Deserializer;
 use serde::{Deserialize, Serialize};
@@ -123,50 +124,6 @@ impl<'de> Deserialize<'de> for EnvironmentStatus {
                     data: value,
                 })
             }
-        }
-    }
-}
-
-/// Serializes an optional int64 in the protobuf-JSON string form the API
-/// uses on the wire, keeping deserialize-then-serialize roundtrips faithful
-/// to captured responses.
-#[allow(clippy::ref_option)] // signature dictated by serde's serialize_with
-fn serialize_string_i64<S>(value: &Option<i64>, serializer: S) -> Result<S::Ok, S::Error>
-where
-    S: serde::Serializer,
-{
-    match value {
-        Some(n) => serializer.serialize_str(&n.to_string()),
-        None => serializer.serialize_none(),
-    }
-}
-
-/// Deserializes an optional int64 that the API serializes as a JSON string
-/// (protobuf JSON convention), accepting a plain number too.
-fn deserialize_string_i64<'de, D>(deserializer: D) -> Result<Option<i64>, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    let value = Option::<serde_json::Value>::deserialize(deserializer)?;
-    match value {
-        None | Some(serde_json::Value::Null) => Ok(None),
-        Some(serde_json::Value::Number(n)) => {
-            let parsed = n.as_i64();
-            if parsed.is_none() {
-                tracing::warn!("Non-i64 JSON number for int64 field, dropping: {n}");
-            }
-            Ok(parsed)
-        }
-        Some(serde_json::Value::String(s)) => {
-            let parsed = s.parse().ok();
-            if parsed.is_none() {
-                tracing::warn!("Unparseable int64 string from API, dropping: {s:?}");
-            }
-            Ok(parsed)
-        }
-        Some(other) => {
-            tracing::warn!("Unexpected JSON type for int64 field, dropping: {other:?}");
-            Ok(None)
         }
     }
 }
