+++ src/lib.rs
@@ -56,6 +56,7 @@
 // Internal HTTP Layer (pub(crate))
 // =============================================================================
 pub(crate) mod http;
+pub(crate) mod serde_util;
 
 // =============================================================================
 // Core Type Modules
